removes UserContext, introduces a provider instead
This change removes the UserContext object that was being passed to every call so that one application could
manage multiple user ids to Atmos. Instead it has been replaced with the AuthenticationCredentialProvider, which
can look up user ids and shared secrets if need be. Hopefully, this will provide enough flexibility for custom
implementation to resolve what they need.

This should work for many applications that would likely be resolving based on the principle executing the current
thread.